@@ -1,6 +1,6 @@
-const morgan = require('morgan');
-const config = require('./config');
-const logger = require('./logger');
+import morgan from 'morgan';
+import config from './config.js';
+import logger from './logger.js';
 
 morgan.token('message', (req, res) => res.locals.errorMessage || '');
 
@@ -9,16 +9,16 @@ const successResponseFormat = `${getIpFormat()}:method :url :status - :response-
 const errorResponseFormat = `${getIpFormat()}:method :url :status - :response-time ms - message: :message`;
 
 const successHandler = morgan(successResponseFormat, {
-  skip: (req, res) => res.statusCode >= 400,
-  stream: { write: (message) => logger.info(message.trim()) },
+    skip: (req, res) => res.statusCode >= 400,
+    stream: { write: (message) => logger.info(message.trim()) },
 });
 
 const errorHandler = morgan(errorResponseFormat, {
-  skip: (req, res) => res.statusCode < 400,
-  stream: { write: (message) => logger.error(message.trim()) },
+    skip: (req, res) => res.statusCode < 400,
+    stream: { write: (message) => logger.error(message.trim()) },
 });
 
-module.exports = {
-  successHandler,
-  errorHandler,
+export {
+    successHandler,
+    errorHandler,
 };
