@@ -1,30 +1,30 @@
-const { Strategy: JwtStrategy, ExtractJwt } = require('passport-jwt');
-const config = require('./config');
-const { tokenTypes } = require('./tokens');
-const { User } = require('../models');
+import { Strategy as JwtStrategy, ExtractJwt } from 'passport-jwt';
+import config from './config.js';
+import { tokenTypes } from './tokens.js';
+import { User } from '../models/index.js'; // Assuming 'index.js' re-exports User
 
 const jwtOptions = {
-  secretOrKey: config.jwt.secret,
-  jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+    secretOrKey: config.jwt.secret,
+    jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
 };
 
 const jwtVerify = async (payload, done) => {
-  try {
-    if (payload.type !== tokenTypes.ACCESS) {
-      throw new Error('Invalid token type');
+    try {
+        if (payload.type !== tokenTypes.ACCESS) {
+            throw new Error('Invalid token type');
+        }
+        const user = await User.findById(payload.sub);
+        if (!user) {
+            return done(null, false);
+        }
+        done(null, user);
+    } catch (error) {
+        done(error, false);
     }
-    const user = await User.findById(payload.sub);
-    if (!user) {
-      return done(null, false);
-    }
-    done(null, user);
-  } catch (error) {
-    done(error, false);
-  }
 };
 
 const jwtStrategy = new JwtStrategy(jwtOptions, jwtVerify);
 
-module.exports = {
-  jwtStrategy,
+export {
+    jwtStrategy,
 };
