@@ -1,21 +1,21 @@
-const Joi = require('joi');
-const httpStatus = require('http-status');
-const pick = require('../utils/pick');
-const ApiError = require('../utils/ApiError');
+import Joi from 'joi';
+import httpStatus from 'http-status';
+import { pick } from '../utils/pick.js';
+import ApiError from '../utils/ApiError.js';
 
 const validate = (schema) => (req, res, next) => {
-  const validSchema = pick(schema, ['params', 'query', 'body']);
-  const object = pick(req, Object.keys(validSchema));
-  const { value, error } = Joi.compile(validSchema)
-    .prefs({ errors: { label: 'key' }, abortEarly: false })
-    .validate(object);
+    const validSchema = pick(schema, ['params', 'query', 'body']);
+    const object = pick(req, Object.keys(validSchema));
+    const { value, error } = Joi.compile(validSchema)
+        .prefs({ errors: { label: 'key' }, abortEarly: false })
+        .validate(object);
 
-  if (error) {
-    const errorMessage = error.details.map((details) => details.message).join(', ');
-    return next(new ApiError(httpStatus.BAD_REQUEST, errorMessage));
-  }
-  Object.assign(req, value);
-  return next();
+    if (error) {
+        const errorMessage = error.details.map((details) => details.message).join(', ');
+        return next(new ApiError(httpStatus.BAD_REQUEST, errorMessage));
+    }
+    Object.assign(req, value);
+    return next();
 };
 
-module.exports = validate;
+export default validate;
