@@ -1,7 +1,7 @@
 /* eslint-disable no-param-reassign */
 
 const paginate = (schema) => {
-  /**
+    /**
    * @typedef {Object} QueryResult
    * @property {Document[]} results - Results found
    * @property {number} page - Current page
@@ -9,7 +9,7 @@ const paginate = (schema) => {
    * @property {number} totalPages - Total number of pages
    * @property {number} totalResults - Total number of documents
    */
-  /**
+    /**
    * Query for documents with pagination
    * @param {Object} [filter] - Mongo filter
    * @param {Object} [options] - Query options
@@ -19,52 +19,52 @@ const paginate = (schema) => {
    * @param {number} [options.page] - Current page (default = 1)
    * @returns {Promise<QueryResult>}
    */
-  schema.statics.paginate = async function (filter, options) {
-    let sort = '';
-    if (options.sortBy) {
-      const sortingCriteria = [];
-      options.sortBy.split(',').forEach((sortOption) => {
-        const [key, order] = sortOption.split(':');
-        sortingCriteria.push((order === 'desc' ? '-' : '') + key);
-      });
-      sort = sortingCriteria.join(' ');
-    } else {
-      sort = 'createdAt';
-    }
+    schema.statics.paginate = async function (filter, options) {
+        let sort = '';
+        if (options.sortBy) {
+            const sortingCriteria = [];
+            options.sortBy.split(',').forEach((sortOption) => {
+                const [key, order] = sortOption.split(':');
+                sortingCriteria.push((order === 'desc' ? '-' : '') + key);
+            });
+            sort = sortingCriteria.join(' ');
+        } else {
+            sort = 'createdAt';
+        }
 
-    const limit = options.limit && parseInt(options.limit, 10) > 0 ? parseInt(options.limit, 10) : 10;
-    const page = options.page && parseInt(options.page, 10) > 0 ? parseInt(options.page, 10) : 1;
-    const skip = (page - 1) * limit;
+        const limit = options.limit && parseInt(options.limit, 10) > 0 ? parseInt(options.limit, 10) : 10;
+        const page = options.page && parseInt(options.page, 10) > 0 ? parseInt(options.page, 10) : 1;
+        const skip = (page - 1) * limit;
 
-    const countPromise = this.countDocuments(filter).exec();
-    let docsPromise = this.find(filter).sort(sort).skip(skip).limit(limit);
+        const countPromise = this.countDocuments(filter).exec();
+        let docsPromise = this.find(filter).sort(sort).skip(skip).limit(limit);
 
-    if (options.populate) {
-      options.populate.split(',').forEach((populateOption) => {
-        docsPromise = docsPromise.populate(
-          populateOption
-            .split('.')
-            .reverse()
-            .reduce((a, b) => ({ path: b, populate: a }))
-        );
-      });
-    }
+        if (options.populate) {
+            options.populate.split(',').forEach((populateOption) => {
+                docsPromise = docsPromise.populate(
+                    populateOption
+                        .split('.')
+                        .reverse()
+                        .reduce((a, b) => ({ path: b, populate: a }))
+                );
+            });
+        }
 
-    docsPromise = docsPromise.exec();
+        docsPromise = docsPromise.exec();
 
-    return Promise.all([countPromise, docsPromise]).then((values) => {
-      const [totalResults, results] = values;
-      const totalPages = Math.ceil(totalResults / limit);
-      const result = {
-        results,
-        page,
-        limit,
-        totalPages,
-        totalResults,
-      };
-      return Promise.resolve(result);
-    });
-  };
+        return await ([countPromise, docsPromise]).then((values) => {
+            const [totalResults, results] = values;
+            const totalPages = Math.ceil(totalResults / limit);
+            const result = {
+                results,
+                page,
+                limit,
+                totalPages,
+                totalResults,
+            };
+            return (result);
+        });
+    };
 };
 
-module.exports = paginate;
+export default paginate;
