@@ -1,8 +1,8 @@
-const express = require('express');
-const validate = require('../../middlewares/validate');
-const authValidation = require('../../validations/auth.validation');
-const authController = require('../../controllers/auth.controller');
-const auth = require('../../middlewares/auth');
+import express from 'express';
+import { validate } from '../../middlewares/validate';
+import { authValidation } from '../../validations/auth.validation';
+import { authController } from '../../controllers/auth.controller';
+import { auth } from '../../middlewares/auth';
 
 const router = express.Router();
 
@@ -15,7 +15,7 @@ router.post('/reset-password', validate(authValidation.resetPassword), authContr
 router.post('/send-verification-email', auth(), authController.sendVerificationEmail);
 router.post('/verify-email', validate(authValidation.verifyEmail), authController.verifyEmail);
 
-module.exports = router;
+export default router;
 
 /**
  * @swagger
