@@ -1,39 +1,39 @@
-const express = require('express');
-const authRoute = require('./auth.route');
-const userRoute = require('./user.route');
-const docsRoute = require('./docs.route');
-const config = require('../../config/config');
+import express from 'express';
+import authRoute from './auth.route.js';
+import userRoute from './user.route.js';
+import docsRoute from './docs.route.js';
+import config from '../../config/config.js';
 
 const router = express.Router();
 
 const defaultRoutes = [
-  {
-    path: '/auth',
-    route: authRoute,
-  },
-  {
-    path: '/users',
-    route: userRoute,
-  },
+    {
+        path: '/auth',
+        route: authRoute,
+    },
+    {
+        path: '/users',
+        route: userRoute,
+    },
 ];
 
 const devRoutes = [
-  // routes available only in development mode
-  {
-    path: '/docs',
-    route: docsRoute,
-  },
+    // routes available only in development mode
+    {
+        path: '/docs',
+        route: docsRoute,
+    },
 ];
 
 defaultRoutes.forEach((route) => {
-  router.use(route.path, route.route);
+    router.use(route.path, route.route);
 });
 
 /* istanbul ignore next */
 if (config.env === 'development') {
-  devRoutes.forEach((route) => {
-    router.use(route.path, route.route);
-  });
+    devRoutes.forEach((route) => {
+        router.use(route.path, route.route);
+    });
 }
 
-module.exports = router;
+export default router;
