@@ -1,23 +1,23 @@
-const express = require('express');
-const auth = require('../../middlewares/auth');
-const validate = require('../../middlewares/validate');
-const userValidation = require('../../validations/user.validation');
-const userController = require('../../controllers/user.controller');
+import express from 'express';
+import auth from '../../middlewares/auth.js';
+import validate from '../../middlewares/validate.js';
+import userValidation from '../../validations/user.validation.js';
+import userController from '../../controllers/user.controller.js';
 
 const router = express.Router();
 
 router
-  .route('/')
-  .post(auth('manageUsers'), validate(userValidation.createUser), userController.createUser)
-  .get(auth('getUsers'), validate(userValidation.getUsers), userController.getUsers);
+    .route('/')
+    .post(auth('manageUsers'), validate(userValidation.createUser), userController.createUser)
+    .get(auth('getUsers'), validate(userValidation.getUsers), userController.getUsers);
 
 router
-  .route('/:userId')
-  .get(auth('getUsers'), validate(userValidation.getUser), userController.getUser)
-  .patch(auth('manageUsers'), validate(userValidation.updateUser), userController.updateUser)
-  .delete(auth('manageUsers'), validate(userValidation.deleteUser), userController.deleteUser);
+    .route('/:userId')
+    .get(auth('getUsers'), validate(userValidation.getUser), userController.getUser)
+    .patch(auth('manageUsers'), validate(userValidation.updateUser), userController.updateUser)
+    .delete(auth('manageUsers'), validate(userValidation.deleteUser), userController.deleteUser);
 
-module.exports = router;
+export default router;
 
 /**
  * @swagger
