@@ -1,14 +1,14 @@
-const nodemailer = require('nodemailer');
-const config = require('../config/config');
-const logger = require('../config/logger');
+import nodemailer from 'nodemailer';
+import config from '../config/config.js';
+import logger from '../config/logger.js';
 
 const transport = nodemailer.createTransport(config.email.smtp);
 /* istanbul ignore next */
 if (config.env !== 'test') {
-  transport
-    .verify()
-    .then(() => logger.info('Connected to email server'))
-    .catch(() => logger.warn('Unable to connect to email server. Make sure you have configured the SMTP options in .env'));
+    transport
+        .verify()
+        .then(() => logger.info('Connected to email server'))
+        .catch(() => logger.warn('Unable to connect to email server. Make sure you have configured the SMTP options in .env'));
 }
 
 /**
@@ -19,8 +19,8 @@ if (config.env !== 'test') {
  * @returns {Promise}
  */
 const sendEmail = async (to, subject, text) => {
-  const msg = { from: config.email.from, to, subject, text };
-  await transport.sendMail(msg);
+    const msg = { from: config.email.from, to, subject, text };
+    await transport.sendMail(msg);
 };
 
 /**
@@ -30,13 +30,13 @@ const sendEmail = async (to, subject, text) => {
  * @returns {Promise}
  */
 const sendResetPasswordEmail = async (to, token) => {
-  const subject = 'Reset password';
-  // replace this url with the link to the reset password page of your front-end app
-  const resetPasswordUrl = `http://link-to-app/reset-password?token=${token}`;
-  const text = `Dear user,
+    const subject = 'Reset password';
+    // replace this url with the link to the reset password page of your front-end app
+    const resetPasswordUrl = `http://link-to-app/reset-password?token=${token}`;
+    const text = `Dear user,
 To reset your password, click on this link: ${resetPasswordUrl}
 If you did not request any password resets, then ignore this email.`;
-  await sendEmail(to, subject, text);
+    await sendEmail(to, subject, text);
 };
 
 /**
@@ -46,18 +46,18 @@ If you did not request any password resets, then ignore this email.`;
  * @returns {Promise}
  */
 const sendVerificationEmail = async (to, token) => {
-  const subject = 'Email Verification';
-  // replace this url with the link to the email verification page of your front-end app
-  const verificationEmailUrl = `http://link-to-app/verify-email?token=${token}`;
-  const text = `Dear user,
+    const subject = 'Email Verification';
+    // replace this url with the link to the email verification page of your front-end app
+    const verificationEmailUrl = `http://link-to-app/verify-email?token=${token}`;
+    const text = `Dear user,
 To verify your email, click on this link: ${verificationEmailUrl}
 If you did not create an account, then ignore this email.`;
-  await sendEmail(to, subject, text);
+    await sendEmail(to, subject, text);
 };
 
-module.exports = {
-  transport,
-  sendEmail,
-  sendResetPasswordEmail,
-  sendVerificationEmail,
+export {
+    transport,
+    sendEmail,
+    sendResetPasswordEmail,
+    sendVerificationEmail,
 };
