@@ -1,5 +1,5 @@
 const catchAsync = (fn) => (req, res, next) => {
-  Promise.resolve(fn(req, res, next)).catch((err) => next(err));
+    Promise.resolve(fn(req, res, next)).catch((err) => next(err));
 };
 
-module.exports = catchAsync;
+export default catchAsync;
