@@ -5,13 +5,13 @@
  * @returns {Object}
  */
 const pick = (object, keys) => {
-  return keys.reduce((obj, key) => {
-    if (object && Object.prototype.hasOwnProperty.call(object, key)) {
-      // eslint-disable-next-line no-param-reassign
-      obj[key] = object[key];
-    }
-    return obj;
-  }, {});
+    return keys.reduce((obj, key) => {
+        if (object && Object.prototype.hasOwnProperty.call(object, key)) {
+            // eslint-disable-next-line no-param-reassign
+            obj[key] = object[key];
+        }
+        return obj;
+    }, {});
 };
 
-module.exports = pick;
+export default pick;
