@@ -1,60 +1,60 @@
-const Joi = require('joi');
-const { password } = require('./custom.validation');
+import Joi from 'joi';
+import { password } from './custom.validation.js';
 
 const register = {
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required().custom(password),
-    name: Joi.string().required(),
-  }),
+    body: Joi.object().keys({
+        email: Joi.string().required().email(),
+        password: Joi.string().required().custom(password),
+        name: Joi.string().required(),
+    }),
 };
 
 const login = {
-  body: Joi.object().keys({
-    email: Joi.string().required(),
-    password: Joi.string().required(),
-  }),
+    body: Joi.object().keys({
+        email: Joi.string().required(),
+        password: Joi.string().required(),
+    }),
 };
 
 const logout = {
-  body: Joi.object().keys({
-    refreshToken: Joi.string().required(),
-  }),
+    body: Joi.object().keys({
+        refreshToken: Joi.string().required(),
+    }),
 };
 
 const refreshTokens = {
-  body: Joi.object().keys({
-    refreshToken: Joi.string().required(),
-  }),
+    body: Joi.object().keys({
+        refreshToken: Joi.string().required(),
+    }),
 };
 
 const forgotPassword = {
-  body: Joi.object().keys({
-    email: Joi.string().email().required(),
-  }),
+    body: Joi.object().keys({
+        email: Joi.string().email().required(),
+    }),
 };
 
 const resetPassword = {
-  query: Joi.object().keys({
-    token: Joi.string().required(),
-  }),
-  body: Joi.object().keys({
-    password: Joi.string().required().custom(password),
-  }),
+    query: Joi.object().keys({
+        token: Joi.string().required(),
+    }),
+    body: Joi.object().keys({
+        password: Joi.string().required().custom(password),
+    }),
 };
 
 const verifyEmail = {
-  query: Joi.object().keys({
-    token: Joi.string().required(),
-  }),
+    query: Joi.object().keys({
+        token: Joi.string().required(),
+    }),
 };
 
-module.exports = {
-  register,
-  login,
-  logout,
-  refreshTokens,
-  forgotPassword,
-  resetPassword,
-  verifyEmail,
+export {
+    register,
+    login,
+    logout,
+    refreshTokens,
+    forgotPassword,
+    resetPassword,
+    verifyEmail,
 };
