@@ -1,54 +1,54 @@
-const Joi = require('joi');
-const { password, objectId } = require('./custom.validation');
+import Joi from 'joi';
+import { password, objectId } from './custom.validation.js';
 
 const createUser = {
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required().custom(password),
-    name: Joi.string().required(),
-    role: Joi.string().required().valid('user', 'admin'),
-  }),
+    body: Joi.object().keys({
+        email: Joi.string().required().email(),
+        password: Joi.string().required().custom(password),
+        name: Joi.string().required(),
+        role: Joi.string().required().valid('user', 'admin'),
+    }),
 };
 
 const getUsers = {
-  query: Joi.object().keys({
-    name: Joi.string(),
-    role: Joi.string(),
-    sortBy: Joi.string(),
-    limit: Joi.number().integer(),
-    page: Joi.number().integer(),
-  }),
+    query: Joi.object().keys({
+        name: Joi.string(),
+        role: Joi.string(),
+        sortBy: Joi.string(),
+        limit: Joi.number().integer(),
+        page: Joi.number().integer(),
+    }),
 };
 
 const getUser = {
-  params: Joi.object().keys({
-    userId: Joi.string().custom(objectId),
-  }),
+    params: Joi.object().keys({
+        userId: Joi.string().custom(objectId),
+    }),
 };
 
 const updateUser = {
-  params: Joi.object().keys({
-    userId: Joi.required().custom(objectId),
-  }),
-  body: Joi.object()
-    .keys({
-      email: Joi.string().email(),
-      password: Joi.string().custom(password),
-      name: Joi.string(),
-    })
-    .min(1),
+    params: Joi.object().keys({
+        userId: Joi.required().custom(objectId),
+    }),
+    body: Joi.object()
+        .keys({
+            email: Joi.string().email(),
+            password: Joi.string().custom(password),
+            name: Joi.string(),
+        })
+        .min(1),
 };
 
 const deleteUser = {
-  params: Joi.object().keys({
-    userId: Joi.string().custom(objectId),
-  }),
+    params: Joi.object().keys({
+        userId: Joi.string().custom(objectId),
+    }),
 };
 
-module.exports = {
-  createUser,
-  getUsers,
-  getUser,
-  updateUser,
-  deleteUser,
+export {
+    createUser,
+    getUsers,
+    getUser,
+    updateUser,
+    deleteUser,
 };
