@@ -1,14 +1,14 @@
-const moment = require('moment');
-const config = require('../../src/config/config');
-const { tokenTypes } = require('../../src/config/tokens');
-const tokenService = require('../../src/services/token.service');
-const { userOne, admin } = require('./user.fixture');
+import moment from 'moment';
+import config from '../../src/config/config.js';
+import { tokenTypes } from '../../src/config/tokens.js';
+import tokenService from '../../src/services/token.service.js';
+import { userOne, admin } from './user.fixture.js';
 
 const accessTokenExpires = moment().add(config.jwt.accessExpirationMinutes, 'minutes');
 const userOneAccessToken = tokenService.generateToken(userOne._id, accessTokenExpires, tokenTypes.ACCESS);
 const adminAccessToken = tokenService.generateToken(admin._id, accessTokenExpires, tokenTypes.ACCESS);
 
-module.exports = {
-  userOneAccessToken,
-  adminAccessToken,
+export {
+    userOneAccessToken,
+    adminAccessToken,
 };
