@@ -1,18 +1,18 @@
-const mongoose = require('mongoose');
-const config = require('../../src/config/config');
+import mongoose from 'mongoose';
+import config from '../../src/config/config.js';
 
 const setupTestDB = () => {
-  beforeAll(async () => {
-    await mongoose.connect(config.mongoose.url, config.mongoose.options);
-  });
+    beforeAll(async () => {
+        await mongoose.connect(config.mongoose.url, config.mongoose.options);
+    });
 
-  beforeEach(async () => {
-    await Promise.all(Object.values(mongoose.connection.collections).map(async (collection) => collection.deleteMany()));
-  });
+    beforeEach(async () => {
+        await Promise.all(Object.values(mongoose.connection.collections).map(async (collection) => collection.deleteMany()));
+    });
 
-  afterAll(async () => {
-    await mongoose.disconnect();
-  });
+    afterAll(async () => {
+        await mongoose.disconnect();
+    });
 };
 
-module.exports = setupTestDB;
+export default setupTestDB;
