@@ -1,12 +1,12 @@
 const allRoles = {
-  user: [],
-  admin: ['getUsers', 'manageUsers'],
+    user: [],
+    admin: ['getUsers', 'manageUsers'],
 };
 
 const roles = Object.keys(allRoles);
 const roleRights = new Map(Object.entries(allRoles));
 
-module.exports = {
-  roles,
-  roleRights,
+export {
+    roles,
+    roleRights,
 };
