@@ -1,2 +1,2 @@
-module.exports.authController = require('./auth.controller');
-module.exports.userController = require('./user.controller');
+export { default as authController } from './auth.controller';
+export { default as userController } from './user.controller';
