@@ -1,2 +1,2 @@
-module.exports.Token = require('./token.model');
-module.exports.User = require('./user.model');
+export { default as Token } from './token.model';
+export { default as User } from './user.model';
