@@ -1,4 +1,4 @@
-module.exports.authService = require('./auth.service');
-module.exports.emailService = require('./email.service');
-module.exports.tokenService = require('./token.service');
-module.exports.userService = require('./user.service');
+export { authService } from './auth.service';
+export { emailService } from './email.service';
+export { tokenService } from './token.service';
+export { userService } from './user.service';
