@@ -1,2 +1,2 @@
-module.exports.authValidation = require('./auth.validation');
-module.exports.userValidation = require('./user.validation');
+export { default as authValidation } from './auth.validation';
+export { default as userValidation } from './user.validation';
